Cookies

+++ app/controllers/User.php
@@ -6,11 +6,29 @@ class User extends Controller {
     Middleware::auth();
   }
 
+  public function add() {
+    if (!empty($_POST)) {
+      if($this->model("Users")->validate()) {
+        Flasher::setFlash("Username sudah digunakan!", "warning");
+      } else {
+        $this->model("Users")->insert();
+        Flasher::setFlash("Data User telah ditambahkan!", "success");
+        return Functions::redirect("dashboard/user");
+      }
+    }
+    Functions::redirect("dashboard/user");
+  }
+
   public function remove($id = null) {
     if ($id) {
-      $this->model("Users")->delete($id);
-      Flasher::setFlash("Berhasil menghapus data user", "success");
-      Functions::redirect("dashboard/user");
+      if ($id == $_SESSION['user']['id']) {
+        Flasher::setFlash("Tidak bisa menghapus data sendiri!", "warning");
+        Functions::redirect("dashboard/user");
+      } else {
+        $this->model("Users")->delete($id);
+        Flasher::setFlash("Berhasil menghapus data user", "success");
+        Functions::redirect("dashboard/user");
+      }
     }
     Functions::redirect("dashboard/user");
   }
